ST-645 PMD fix
