+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/util/DocumentListUtil.java
@@ -20,12 +20,8 @@ import java.util.stream.Collectors;
 
 import static uk.gov.hmcts.sptribs.caseworker.util.DecisionDocumentListUtil.getDecisionDocs;
 import static uk.gov.hmcts.sptribs.caseworker.util.DecisionDocumentListUtil.getFinalDecisionDocs;
-import static uk.gov.hmcts.sptribs.caseworker.util.DecisionDocumentListUtil.removeDecisionDoc;
-import static uk.gov.hmcts.sptribs.caseworker.util.DecisionDocumentListUtil.removeFinalDecisionDoc;
 import static uk.gov.hmcts.sptribs.caseworker.util.DocumentManagementUtil.buildListValues;
-import static uk.gov.hmcts.sptribs.caseworker.util.DocumentManagementUtil.checkLists;
 import static uk.gov.hmcts.sptribs.caseworker.util.OrderDocumentListUtil.getOrderDocuments;
-import static uk.gov.hmcts.sptribs.caseworker.util.OrderDocumentListUtil.removeOrderDoc;
 
 public final class DocumentListUtil {
     private DocumentListUtil() {
@@ -148,39 +144,5 @@ public final class DocumentListUtil {
         return buildListValues(getOrderDocuments(cicCase));
     }
 
-    public static CaseData removeEvaluatedListDoc(CaseData caseData, CaseData oldData) {
-        removeDecisionDoc(caseData, oldData);
-        removeFinalDecisionDoc(caseData, oldData);
-        var cic = caseData.getCicCase();
-        removeOrderDoc(cic, oldData.getCicCase());
-        if (!CollectionUtils.isEmpty(oldData.getDocManagement().getCaseworkerCICDocument())
-            && (CollectionUtils.isEmpty(caseData.getDocManagement().getCaseworkerCICDocument())
-            || caseData.getDocManagement().getCaseworkerCICDocument().size()
-            < oldData.getDocManagement().getCaseworkerCICDocument().size())) {
-            checkLists(caseData, oldData.getDocManagement().getCaseworkerCICDocument(),
-                caseData.getDocManagement().getCaseworkerCICDocument());
-        }
-        if (!CollectionUtils.isEmpty(oldData.getCloseCase().getDocuments())
-            && (CollectionUtils.isEmpty(caseData.getCloseCase().getDocuments())
-            || caseData.getCloseCase().getDocuments().size() < oldData.getCloseCase().getDocuments().size())) {
-            checkLists(caseData, oldData.getCloseCase().getDocuments(), caseData.getCloseCase().getDocuments());
-        }
-        if (!CollectionUtils.isEmpty(oldData.getCicCase().getReinstateDocuments())
-            && (CollectionUtils.isEmpty(caseData.getCicCase().getReinstateDocuments())
-            || cic.getReinstateDocuments().size() < oldData.getCicCase().getReinstateDocuments().size())) {
-            checkLists(caseData, oldData.getCicCase().getReinstateDocuments(), cic.getReinstateDocuments());
-        }
-        if (!CollectionUtils.isEmpty(oldData.getCicCase().getApplicantDocumentsUploaded())
-            && (CollectionUtils.isEmpty(cic.getApplicantDocumentsUploaded())
-            || cic.getApplicantDocumentsUploaded().size() < oldData.getCicCase().getApplicantDocumentsUploaded().size())) {
-            checkLists(caseData, oldData.getCicCase().getApplicantDocumentsUploaded(), cic.getApplicantDocumentsUploaded());
-        }
-        if (!CollectionUtils.isEmpty(oldData.getListing().getSummary().getRecFile())
-            && (CollectionUtils.isEmpty(caseData.getListing().getSummary().getRecFile())
-            || caseData.getListing().getSummary().getRecFile().size() < oldData.getListing().getSummary().getRecFile().size())) {
-            checkLists(caseData, oldData.getListing().getSummary().getRecFile(), caseData.getListing().getSummary().getRecFile());
-        }
-        return caseData;
-    }
 
 }
